+++ lib/tufy/build_header_segment.rb
@@ -1,16 +1,15 @@
 module Tufy
   class BuildHeaderSegment < BuildSegment
-    expects :raw_data
+    expects :header_data
     promises :transformed_data
 
     REQUIRED_KEYS = [
-      :member_reference_number,
       :processor_name,
       :user_id
     ]
 
     executed do |ctx|
-      validate_presence_of_required_keys(ctx.raw_data, REQUIRED_KEYS)
+      validate_presence_of_required_keys(ctx.header_data, REQUIRED_KEYS)
       ctx.transformed_data = ctx.transformed_data + transform(ctx).upcase
     end
 
@@ -19,9 +18,9 @@ module Tufy
     def self.transform(ctx)
       Constants::SEGMENT_TAG + # Segment Tag (Required)
         Constants::TUDF_VERSION + # Version Tag (Required)
-        "#{FormatStrings::F25TS % ctx.raw_data[:member_reference_number]}" + # Member Reference Number (Required)
-        "#{FormatStrings::F26TS % ctx.raw_data[:processor_name].upcase}" + # Member Processor Name (Required)
-        "#{FormatStrings::F10TS % ctx.raw_data[:user_id].upcase}" + # User ID (Required)
+        "#{FormatStrings::F25TS % Time.now.to_i.to_s}" + # Member Reference Number (Required)
+        "#{FormatStrings::F26TS % ctx.header_data[:processor_name].upcase}" + # Member Processor Name (Required)
+        "#{FormatStrings::F10TS % ctx.header_data[:user_id].upcase}" + # User ID (Required)
         "#{transform_date(Date.today)}" + # Reported Date (Required)
         Constants::FILLER # Filler (Required)
     end