+++ lib/tufy/transform_raw_data.rb
@@ -1,0 +1,8 @@
+module Tufy
+  class TransformRawData
+
+    def self.execute(ctx)
+      BuildRecord.execute(ctx)
+    end
+  end
+end