+++ lib/tufy/transform_raw_data_array.rb
@@ -1,0 +1,14 @@
+module Tufy
+  class TransformRawDataArray
+    include LightService::Organizer
+
+    def self.execute(ctx)
+      ctx[:transformed_data] = ''
+      ctx = with(ctx).reduce(
+        BuildHeaderSegment,
+        BuildRecords,
+      )
+      ctx[:transformed_data]
+    end
+  end
+end