+++ CHANGELOG.md
@@ -1,4 +1,10 @@
-# v0.05
+# v0.0.6
+
+- Added Tufy.transform_raw_data_array and refactor Tufy.transform
+- Renamed Tufy::BuildAccountSegment.tu_past_due_code to ::past_due_code
+- Use Unix timestamp for the Member Reference Number in the Header Segment
+
+# v0.0.5
 
 - Create method to return TU past due codes instead of using constants